Adding dev sprints to schedule page.
To alleviate some confusion that I've seen out there about why the schedule
page only lists the 9th-11th when the front page says the conference runs
from the 9th to the 13th, I'd added some of the text from last year (from
PyCon US originally, I think?) to the stub page for the conference schedule.

In the longer term, this info should likely move to a more permanent home,
like a dedicated "sprints" page -- but the dates should still be on the
schedule along with some basic info.