Adds a manage.py script to assign reviewers to talks.
Requires that a 'reviewers' group exist, and that there are people assigned to it (otherwise it won't find any reviewers to add!).

It will assign up to 3 reviewers, where those reviewers are NOT one of the speakers, and that reviewer has not opted out of reviewing that talk. It will choose random reviewers from those with the fewest existing not-opted-out assignments.